Run destructive schema test at the end
